requirements.txt
mss
pytesseract
Pillow
